refactor: simplify data require toggle
