+++ components/header/AvatarButton.vue
@@ -8,6 +8,12 @@ const { signIn, signOut, status, data } = useAuth();
 
 const menu = ref();
 
+const { datastoreState, setDatastoreRequired } =
+  await useDatastoreRequirement();
+const datastoreRequired = computed(
+  () => datastoreState.value.datastoreRequired,
+);
+
 const config = useRuntimeConfig();
 const baseUrl = new URL(config.public.baseUrl).origin;
 const idpProvider = config.public.idpProvider;
@@ -28,6 +34,7 @@ const menuItems = ref([
         label: userActionLabel,
         icon: userActionIcon,
         command: () => {
+          // eslint-disable-next-line @typescript-eslint/no-unused-expressions
           isAuthenticated.value ? signOut() : signIn(`${idpProvider}`);
         },
       },
@@ -44,6 +51,11 @@ const menuItems = ref([
           showCleanupDialog.value = true;
         },
         disabled: !isAuthenticated.value,
+      },
+      {
+        label: "Require Data Store",
+        icon: "pi pi-database",
+        isToggle: true, // custom flag to identify it in the slot
       },
     ],
   },
@@ -92,6 +104,22 @@ const toggle = (event) => {
     >
       <template #item="{ item, props }">
         <a
+          v-if="item.isToggle"
+          v-ripple
+          class="flex items-center"
+          v-bind="props.action"
+          @click.prevent="setDatastoreRequired(!datastoreRequired)"
+        >
+          <i :class="item.icon" />
+          <span class="ml-2 menu-item-label">{{ item.label }}</span>
+          <ToggleSwitch
+            :modelValue="datastoreRequired ?? false"
+            class="ml-auto"
+            @click.stop="setDatastoreRequired(!datastoreRequired)"
+          />
+        </a>
+        <a
+          v-else
           v-ripple
           :href="item.url"
           :target="item.target"
