+++ components/header/MenuHeader.vue
@@ -1,9 +1,8 @@
 <script lang="ts" setup>
 import Menubar from "primevue/menubar";
-import AvatarButton from "~/components/header/AvatarButton/AvatarButton.vue";
+import AvatarButton from "~/components/header/AvatarButton.vue";
 import DarkModeToggle from "~/components/header/DarkModeToggle.vue";
 import { RouterLink } from "#vue-router";
-import DataRequirementToggle from "~/components/header/AvatarButton/DataRequirementToggle.vue";
 
 const { status } = useAuth();
 
@@ -92,7 +91,6 @@ const items = ref([
       </template>
       <template #end>
         <div class="header-end">
-          <DataRequirementToggle />
           <DarkModeToggle />
           <div class="flex align-items-center gap-2 avatar-button">
             <AvatarButton />
